+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - 添加 API `/api/block/prependBlock` | Add API `/api/block/prependBlock`
 - 添加 API `/api/block/appendBlock` | Add API `/api/block/appendBlock`
 - 添加 API `/api/block/updateBlock` | Add API `/api/block/updateBlock`
+- 添加 API `/api/block/deleteBlock` | Add API `/api/block/deleteBlock`
 
 ## v0.0.5 / 2023-06-30